Commit 10-ToDo: Corrigé erreur d'implementation (Commit09_ToDo).
Le rôle d'AbstractManager n'est pas de produire des DAOs mais
d'embarquer la DaoFactory!!!

TODO: Bean "dataSourceTicket" pour DataSource à implementer
