+++ ticket/ticket-business/src/main/java/org/example/demo/ticket/business/manager/impl/AbstractManager.java
@@ -1,33 +1,19 @@
 package org.example.demo.ticket.business.manager.impl;
 
-import org.example.demo.ticket.consumer.contract.dao.ProjetDao;
-import org.example.demo.ticket.consumer.contract.dao.TicketDao;
+import org.example.demo.ticket.consumer.contract.factory.DaoFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public abstract class AbstractManager {
-	
+
 	@Autowired
-	private TicketDao ticketDao;
-	
-	@Autowired
-	private ProjetDao projetDao;
+	private DaoFactory daoFact;
 
-	protected TicketDao getTicketDao() {
-		return ticketDao;
+	public void setDaoFact(DaoFactory daoFact) {
+		this.daoFact = daoFact;
 	}
 
-	public void setTicketDao(TicketDao ticketDao) {
-		this.ticketDao = ticketDao;
+	protected DaoFactory getDaoFact() {
+		return daoFact;
 	}
-
-	protected ProjetDao getProjetDao() {
-		return projetDao;
-	}
-
-	public void setProjetDao(ProjetDao projetDao) {
-		this.projetDao = projetDao;
-	}
-	
-	
 
 }
